+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this branch will be documented in this file.
 
 - Prevented Quick Access History and Task View settings from interfering with each other #103.
 - Replaced legacy APIs with modern Windows methods in `SystemInfo` page to prevent incorrect behavior #105.
+- Reworked the Debloat logic to use the native Windows Deployment API instead of Powershell for app detection and removal.
 
 ## 1.7.0 - Released